Correct the calculation of same in winklermod
The value of same is not calculated correctly when one string fully matches the leading characters of the other string and is shorter than four characters.  The proposed change fixes the calculation of same for that case.  Please see the notebook at http://nbviewer.ipython.org/github/machinelearningdeveloper/same-demo/blob/master/same_demo.ipynb to see a demonstration of the problem and its solution.